% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{as.data.frame.svis_points}
\alias{as.data.frame.svis_points}
\title{as.data.frame.svis_points}
\usage{
\method{as.data.frame}{svis_points}(x, row.names, optional, ...)
}
\arguments{
\item{x}{A svis_points_object}

\item{row.names}{the row names}

\item{optional}{Optional arguements to the data.frame method}

\item{...}{Other arguments}
}
\value{
A data.frame
}
\description{
as.data.frame.svis_points
}
