% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{as.svis_points}
\alias{as.svis_points}
\title{as.svis_points}
\usage{
as.svis_points(x, ...)
}
\arguments{
\item{x}{another object to be coerced into a svis_points}

\item{...}{Other arguments}
}
\value{
A svis_points object
}
\description{
as.svis_points
}
