% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{as.svis_points.svis_layer}
\alias{as.svis_points.svis_layer}
\title{as.svis_points.svis_layer}
\usage{
\method{as.svis_points}{svis_layer}(x, ...)
}
\arguments{
\item{x}{A svis_layer}

\item{...}{Other arguments}
}
\value{
svis_points
}
\description{
as.svis_points.svis_layer
}
