% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{colfunction}
\alias{colfunction}
\title{colfunction}
\usage{
colfunction(x)
}
\arguments{
\item{x}{A svis_layer}
}
\value{
The name of the getcolour function in the js for the fill colour
}
\description{
colfunction
}
