% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{convert_to_geojson}
\alias{convert_to_geojson}
\title{convert_to_geojson}
\usage{
convert_to_geojson(spatial_object)
}
\arguments{
\item{spatial_object}{A sp::SpatialPointsDataFrame, sp::SpatialLinessDataFrame or sp::SpatialPolygonsDataFrame}
}
\value{
A character vector
}
\description{
convert_to_geojson
}
