% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{convert_to_sppts}
\alias{convert_to_sppts}
\title{convert_to_sppts}
\usage{
convert_to_sppts(df, input_proj, output_proj = WGS84(), long, lat)
}
\arguments{
\item{df}{A data.frame that contains X and Y coordinates}

\item{input_proj}{The projection of the  input}

\item{output_proj}{The projection of the output}

\item{long}{The name of the variable that contains the longitude}

\item{lat}{The name of the variable that contains the latitude}
}
\value{
SpatialPointsDataFrame
}
\description{
Convert a data.frame with X, Y points to a spatial point
data.frame
}
