% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{crs.data.frame}
\alias{crs.data.frame}
\title{crs.data.frame}
\usage{
\method{crs}{data.frame}(data)
}
\arguments{
\item{data}{a data.frame with x and y}
}
\value{
one of c("sweref99", "RT90", "WGS84")
}
\description{
crs.data.frame
}
