% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{crs.SpatialPointsDataFrame}
\alias{crs.SpatialPointsDataFrame}
\title{crs.SpatialPointsDataFrame}
\usage{
\method{crs}{SpatialPointsDataFrame}(x, ...)
}
\arguments{
\item{x}{A SpatialPointsDataFrame}

\item{...}{other arguements}
}
\value{
One of these c("sweref99", "RT90", "WGS84")
}
\description{
crs.SpatialPointsDataFrame
}
