% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{epsg}
\alias{epsg}
\title{epsg}
\usage{
epsg(x = c("sweref99", "RT90", "WGS84"))
}
\arguments{
\item{x}{A character}

\item{...}{other arguements}
}
\value{
numeric value
}
\description{
The epsg of various crs
}
