% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_parts.R
\name{fillColor_js}
\alias{fillColor_js}
\title{fillColor_js}
\usage{
fillColor_js(layername, values = NULL, col = NULL,
  default_col = "#000000")
}
\arguments{
\item{layername}{the name of the layer that the function should
be applied}

\item{values}{The numeric values of the factor variable to
consider}

\item{col}{The colours to match with each level}

\item{default_col}{The colour given to a point that doesn't meet
any of the criteria}
}
\value{
An hlt script object
}
\description{
fillColor_js
}
