% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{format.svis_layer}
\alias{format.svis_layer}
\title{format.svis_layer}
\usage{
\method{format}{svis_layer}(x, ...)
}
\arguments{
\item{x}{svis_layer}

\item{...}{Other arguments}
}
\description{
format.svis_layer
}
