% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{format.svis_point}
\alias{format.svis_point}
\title{format.svis_point}
\usage{
\method{format}{svis_point}(x, ...)
}
\arguments{
\item{x}{A svis_point}

\item{...}{Other parameters}
}
\value{
formated point in json
}
\description{
format.svis_point
}
