% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{layer}
\alias{layer}
\title{build a layer}
\usage{
layer(data, ...)
}
\arguments{
\item{data}{svis_points}

\item{...}{Arguments passed to specific layer functions for
different classes}
}
\value{
a svis_layer object
}
\description{
build a layer
}
