% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{layer.SpatialPointsDataFrame}
\alias{layer.SpatialPointsDataFrame}
\title{layer}
\usage{
\method{layer}{SpatialPointsDataFrame}(data, layer_title = "layer1",
  byvar = NULL, col = NULL, ...)
}
\arguments{
\item{data}{The dataset in sp::SpatialPointsDataFrame format}

\item{layer_title}{The human readable name of the layer in the map}

\item{byvar}{The variable in the dataframe to use to colour the
points. Will be coerced into a factor. If NULL, one colour
will be used.}

\item{col}{The colour(s) for the points. If you supply less
colours than the number of levels in 'byvar', the colours
will be reused. You my not supply more colours than levels in
'byvar'.}

\item{...}{Other arguments}
}
\value{
A svis_layer
}
\description{
layer
}
