% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{layer.svis_points}
\alias{layer.svis_points}
\title{layer}
\usage{
\method{layer}{svis_points}(data, layer_title = "layer1", byvar = NULL,
  col = NULL, outline_col = NULL, radius = NULL,
  fillOpacity = NULL, opacity = NULL, ...)
}
\arguments{
\item{data}{The dataset in svis_points format}

\item{layer_title}{The human readable name of the layer in the map}

\item{byvar}{The variable in the dataframe to use to colour the
points. Will be coerced into a factor. If NULL, one colour
will be used.}

\item{col}{The colour(s) for the points. If you supply less
colours than the number of levels in 'byvar', the colours will
be reused. You my not supply more colours than levels in
'byvar'.}

\item{outline_col}{the Outline colour of th points}

\item{radius}{The radius of the points}

\item{fillOpacity}{The fill opacity of the points}

\item{opacity}{The opacity of the opline of the points}

\item{...}{Other arguments}
}
\value{
A svis_layer
}
\description{
layer
}
