% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_div}
\alias{map_div}
\title{map_div}
\usage{
map_div(layers, layer_states = TRUE, center = c(63, 17.5), zoom = 5,
  maxBounds = list(c(54, 4), c(70, 31)))
}
\arguments{
\item{layers}{A layer or a list of layers}

\item{layer_states}{a logical vector indicating the default on/off
state of each layer}

\item{center}{The center of the map}

\item{zoom}{The default zoom level}

\item{maxBounds}{The max navigable bounds of the map. Must be a
list of length 2 with the bottom left corner and top right
corner.}
}
\value{
A svis_div object
}
\description{
map_div
}
