% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_parts.R
\name{maplegend.svis_layer}
\alias{maplegend.svis_layer}
\title{maplegend}
\usage{
\method{maplegend}{svis_layer}(x, position = "bottomright")
}
\arguments{
\item{x}{a svis_layer}
}
\value{
an hlt script
}
\description{
maplegend
}
