% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_parts.R
\name{maplegend.svis_layers}
\alias{maplegend.svis_layers}
\title{maplegend}
\usage{
\method{maplegend}{svis_layers}(x)
}
\arguments{
\item{x}{a svis_layers}
}
\value{
an hlt script
}
\description{
maplegend
}
