% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{names.svis_layer}
\alias{names.svis_layer}
\title{names.svis_layer}
\usage{
\method{names}{svis_layer}(x)
}
\arguments{
\item{x}{a svis_layer object}
}
\value{
name of the layer
}
\description{
names.svis_layer
}
