% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_parts.R
\name{onEachFeature}
\alias{onEachFeature}
\title{onEachFeature}
\usage{
onEachFeature(name = "onEachPoint")
}
\arguments{
\item{name}{name of the function}
}
\value{
A character vector
}
\description{
onEachFeature
}
