% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{opacity}
\alias{opacity}
\title{opacity}
\usage{
opacity(x)
}
\arguments{
\item{x}{A svis_layer}
}
\value{
The opacity of the points of the name of the variable in
    the data to use to determine the opacity
}
\description{
opacity
}
