% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{outline_color}
\alias{outline_color}
\title{outline_color}
\usage{
outline_color(x)
}
\arguments{
\item{x}{A svis_layer}
}
\value{
The outline colour of the points or the name of the variable in
    the data to use to determine the outline colour
}
\description{
outline_color
}
