% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{outline_opacity}
\alias{outline_opacity}
\title{outline_opacity}
\usage{
outline_opacity(x)
}
\arguments{
\item{x}{A svis_layer}
}
\value{
The outline opacity of the points or the name of the variable in
    the data to use to determine the outline opacity
}
\description{
outline_opacity
}
