% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{overlays}
\alias{overlays}
\title{Overlay functions}
\usage{
overlays(object)
}
\arguments{
\item{object}{A svis_layer or svis_layers object}
}
\value{
character the overlay string
}
\description{
Overlay functions
}
