% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{overlays.svis_layer}
\alias{overlays.svis_layer}
\title{overlays.svis_layer}
\usage{
\method{overlays}{svis_layer}(object)
}
\arguments{
\item{object}{a svis_layer}
}
\value{
chracter the overlay string
}
\description{
overlays.svis_layer
}
