% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{overlays.svis_layers}
\alias{overlays.svis_layers}
\title{overlays.svis_layers}
\usage{
\method{overlays}{svis_layers}(object)
}
\arguments{
\item{object}{a svis_layers object}
}
\value{
html snippet
}
\description{
overlays.svis_layers
}
