% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page_header}
\alias{page_header}
\title{page_header}
\usage{
page_header(title = "map")
}
\arguments{
\item{title}{The title of the page}
}
\value{
An hlt object with the page header
}
\description{
Create some of the content in the header of the webpage with the
leaflet map
}
