% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page_page}
\alias{page_page}
\title{page_page}
\usage{
page_page(header, body)
}
\arguments{
\item{header}{an html header}

\item{body}{an html body}
}
\value{
an hlt page
}
\description{
page_page
}
\examples{
\dontrun{

library(svis)

# Get the sample data
pts <- sample_data()
a <- convert_to_geojson(pts)

# Convert to layer objects with 2 different names:
layer1 <- point_layer(a)
layer2 <- point_layer(a, layer_title = "layer2")

# Make a layers object with both layers
layersob <- layers(list(layer1, layer2))

# Generate the page header
header <- page_header()

# Map with a single layer:
body <- page_body(map_div(layer1))
page <- page_page(header, body)
map <- tempfile(fileext = ".html")
capture.output(page, file = map)
browseURL(map)

# Map with 2 layers:

body <- page_body(map_div(layersob))
page <- page_page(header, body)
map <- tempfile(fileext = ".html")
capture.output(page, file = map)
browseURL(map)
}
}
