% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{point_layer}
\alias{point_layer}
\title{point_layer}
\usage{
point_layer(jsondata, layer_title = "layer1", radius = 5,
  fillColor = shQuote("#00A9CE"), color = shQuote("black"),
  weight = 1, opacity = 1, fillOpacity = 1,
  onEach = onEachFeature())
}
\arguments{
\item{jsondata}{Name of the .js file and the name of the variable
in the file}

\item{layer_title}{The human readable name of the layer in the map}

\item{radius}{Radius of the points (maybe a .js function call)}

\item{fillColor}{The colour to fill the points (maybe a .js
function call)}

\item{color}{The border colour of the points (maybe a .js function
call)}

\item{weight}{The weight, actually I don't know what this
parameter does but.. (maybe a .js function call)}

\item{opacity}{The opacity of the border of the points (maybe a
.js function call)}

\item{fillOpacity}{The opacity of the fill (maybe a .js function
call)}

\item{onEach}{The name of the function to call on each feature}
}
\value{
A svis_layer
}
\description{
point_layer
}
