% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{print.svis_layer}
\alias{print.svis_layer}
\title{print.svis_layer}
\usage{
\method{print}{svis_layer}(x, ...)
}
\arguments{
\item{x}{A svis_layer}

\item{...}{Other arguments}
}
\description{
print.svis_layer
}
