% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{print.svis_point}
\alias{print.svis_point}
\title{print.svis_point}
\usage{
\method{print}{svis_point}(x, ...)
}
\arguments{
\item{x}{A svis_point}

\item{...}{other arguments}
}
\description{
print.svis_point
}
