% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{print.svis_points}
\alias{print.svis_points}
\title{print.svis_points}
\usage{
\method{print}{svis_points}(x, ...)
}
\arguments{
\item{x}{A svis_points object}

\item{...}{Other arguments}
}
\description{
print.svis_points
}
