% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{radius}
\alias{radius}
\title{radius}
\usage{
radius(x)
}
\arguments{
\item{x}{A svis_layer}
}
\value{
The radius of the points or the name of the variable in
    the data to use to determine the radius
}
\description{
radius
}
