% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_sample_data_cwd}
\alias{read_sample_data_cwd}
\title{read_sample_data_cwd}
\usage{
read_sample_data_cwd()
}
\value{
a data.frame
}
\description{
Some dummy sample data to use for a point map (or whatever)
}
