% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{sample_data}
\alias{sample_data}
\title{sample_data}
\usage{
sample_data()
}
\value{
a data.frame
}
\description{
Some dummy sample data to use for a point map (or whatever)
}
\examples{
\dontrun{
# Get the data
df <- sample_data()
str(df)
}
}
