% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{scripts}
\alias{scripts}
\title{Get the scripts}
\usage{
scripts(object)
}
\arguments{
\item{object}{A svis_layer or svis_layers object}
}
\value{
a list of hlt_script objects
}
\description{
Get the scripts
}
