% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{scripts.svis_layer}
\alias{scripts.svis_layer}
\title{scripts.svis_layer}
\usage{
\method{scripts}{svis_layer}(object)
}
\arguments{
\item{object}{A svis_layer object}
}
\value{
a script
}
\description{
scripts.svis_layer
}
