% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{scripts.svis_layers}
\alias{scripts.svis_layers}
\title{scripts.svis_layers}
\usage{
\method{scripts}{svis_layers}(object)
}
\arguments{
\item{object}{a svis_layers object}
}
\value{
a list of scripts
}
\description{
scripts.svis_layers
}
