% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{svis_point}
\alias{svis_point}
\title{svis_point}
\usage{
svis_point(data, id)
}
\arguments{
\item{data}{A data.frame with one row}

\item{id}{The id of the point}
}
\value{
a svis_point
}
\description{
svis_point
}
