% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{svis_points.data.frame}
\alias{svis_points.data.frame}
\title{svis_points.data.frame}
\usage{
\method{svis_points}{data.frame}(data, crs = NULL, ...)
}
\arguments{
\item{data}{A data.frame with at least x and y and columns}

\item{crs}{The crs of the data c("sweref99", "RT90", "WGS84"). If null the function tries to guess the crs.}

\item{...}{Other Arguments}
}
\value{
A svis_points object
}
\description{
svis_points.data.frame
}
