% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{svis_points}
\alias{svis_points}
\title{svis_points}
\usage{
svis_points(data, ...)
}
\arguments{
\item{data}{A dataframe or spatial dataframe}

\item{...}{Other arguments}
}
\value{
A svis_points object
}
\description{
svis_points
}
