% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{svis_points.SpatialPointsDataFrame}
\alias{svis_points.SpatialPointsDataFrame}
\title{svis_points.SpatialPointsDataFrame}
\usage{
\method{svis_points}{SpatialPointsDataFrame}(data, ...)
}
\arguments{
\item{data}{A spatialPointsDataFrame}

\item{...}{Other Arguments}
}
\value{
A svis_points object
}
\description{
svis_points.SpatialPointsDataFrame
}
