% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_parts.R
\name{tiles}
\alias{tiles}
\title{tiles}
\usage{
tiles(topo = "Terräng", streets = "Vägkarta",
  satelite = "Flygfoto")
}
\arguments{
\item{topo}{The name of the topo layer}

\item{streets}{The name of the streets layer}

\item{satelite}{The name of the satelite layer}
}
\value{
an hlt::html_script object
}
\description{
tiles
}
