% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{as.data.frame.svis_points}
\alias{as.data.frame.svis_points}
\title{as.data.frame.svis_points}
\usage{
\method{as.data.frame}{svis_points}(x)
}
\arguments{
\item{x}{A svis_points_object}
}
\value{
A data.frame
}
\description{
as.data.frame.svis_points
}
