% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{format.svis_points}
\alias{format.svis_points}
\title{format.svis_points}
\usage{
\method{format}{svis_points}(x, name = NULL, ...)
}
\arguments{
\item{x}{A svis_points object}

\item{name}{The variable name in the .js}

\item{...}{Other arguments}
}
\value{
formatted object in json
}
\description{
format.svis_points
}
