% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_layer.R
\name{names.svis_layers}
\alias{names.svis_layers}
\title{names.svis_layers}
\usage{
\method{names}{svis_layers}(layers)
}
\arguments{
\item{layer}{a svis_layers object}
}
\value{
names of the layers
}
\description{
names.svis_layers
}
