% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{sample_data}
\alias{sample_data}
\title{sample_data}
\usage{
sample_data(dataset = c("cwd", "asf"))
}
\arguments{
\item{dataset}{The name of the dataset. Either "cwd" or "asf"}
}
\value{
A SpatialPointsDataFrame
}
\description{
Get some sample data
}
\examples{
\dontrun{
# Get the data
library(sp)
pts <- sample_data()
plot(pts)
}
}
